Add new task for generating fake letter responses
We want to mark a letter as being delivered if it's sent using a test
key and the `SEND_LETTERS_ENABLED` flag is False. This was being done
by putting a fake response file in S3, but now that we are removing the
code related to response files we can no longer use this method, so this
adds a task to create a fake callback instead.

The original code to generate fake response files will be removed in a
separate PR in case there are existing tasks for it still in flight when
this is deployed.